Replaced PAPPL patch for using selected ready media for jobs
PAPPL defines a "Default Media" consisting of a paper source, a paper
size, a paper type, and whether it is printed borderless or not. It
also defined the so-called "Ready Media", which are a paper size, a
paper type, and whether to print borderless for each paper tray of the
printer, for the user to be able to tell what is loaded in each tray
(via the "Media" web interface page.

On the "Printing Defaults" web interface page the user selects one of
the trays configured under "Media" as the tray to print from. The
drop-down items tell also what is loaded.

It all looks intuitive but it is not, as if you start the Printer
Application and do mot click the "Save Changes" button in the
"Printing Defaults" it is printed with the "Default Media" which is
nowhere visible in the web interface and not with the "Ready Media"
which you see selected on the "Printing Defaults" page.

This is reported as a bug in PAPPL:

https://github.com/michaelrsweet/pappl/issues/164

Here in  the Printer Application it  was worked around with  a a patch
for  the   papplPrinterSetReadyMedia()  function  which   updated  the
"Default Media" to  the selected (by the tray of  the "Default Media")
"Ready Media". This was not accepted in PAPPL, therefore I switched to
another approach.  On job execution I  do not use the  "Default Media"
but the "Ready Media" of the tray specified in the "Default Media".

This commit switches to the new patch.