"*cupsFilter(2): ..." lines got removed from manufacturer PPDs in error
The global correction of common errors in the manufacturer-supplied,
ready-made PPD files also removed the "*cupsFilter(2): ..."
lines. This was wrong. All included PPD files are non PostScript and
need this line, even many PostScript PPDs need it.

This was probably a copy-and-paste error which slipped in when getting
this scriptlet line from some debian/rules file.

In addition, put some comments into snapcraft.yaml.