In the Snap de-activbate PAPPL's log rotation and allow 256 vendor options
When a job is executed libpappl-retrofit starts several filter
functions and the backend in parallel, in forked processes. PAPPL's
log-rotating mechanism cannot cope with that. When the log file fills
up to the limit, it is filled up in each sub-rocess and rotated in
each sub-process. The quick hammering of several rotation actions
makes the logs getting lost and also logging stop working at all
(probably logging is then sent to file descriptors which are already
closed.

Therefore for the Snap we de-activate log rotation in PAPPL and to not
make the log file fill up the disk we log only with the level ERROR
instead of DEBUG by default. One can easily change the log level via
the web interface if needed.

PAPPL limits the number of vendor-specific printer options/attributes
to 32. This can easily get too small for some busier PPD
files. Therefore we raise this number to 256 when building the Snap's
PAPPL.