Initialize config files for CUPS backends in the server startup script
The configuration files for the CUPS Backends ("snmp", "usb") were
initialized by the general startup script, which is also used when the
Printer Application is started as a client, by a non-root user. Here
the initialization does not work and gives ugly, but harmless error
messages.

Now the initialization of these files is moved to the server startup
script and so the error messages go away.