Build extra filters by ourselves to avoid dependency hell
For the Foomatic PPDs which need filter executables separate from
Ghostscript we had installed the Debian packages of the drivers. This
pulled in a lot of dependencies, like CUPS and Ghostscript, having
these duplicate, once the new version we build in the Snap and the
two-year-old version of Ubuntu Core 20 (= 20.04 LTS).

To avoid this, we now build these drivers by ourselves in the Snap
build.

We also include now the PPD files and printer drivers included with
CUPS and cups-filters.