Added more drivers: c2esp, dymo, foo2zjs
Added the first drivers of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

From foo2zjs the PPDs for printers which need their firmware loaded
every time when they are turned on are removed. An facility to
auto-load the firmware would be too complex for this Snap, as 99% do
no need to load firmware. Also the places from which foo2zjs loads
firmware can quickly go away, especially as it is not legal for third
parties to make the non-free firmware files available.

Therefore it is recommended to use the HPLIP Printer Aplication for
these printers. It uses the official proprietary plugin from HP,
downloaded from the official sources, and integrity-verified by
checksum and digital signature.

Color profiles are supported by the included foo2zjs though. they
can be put into appropriate directories in

    /var/snap/ghostscript-printer-app/common/foo2zjs/

Otherwise using these drivers is as simple as adding a printer and
choosing either auto-selection or the appropriate model from the
driver list.