c2050 driver: Small code fix to avoid crash of the driver filter
The simple patch (done inline with Perl) silences the warnings of
implicit declarations as int and avoids a buffer overflow crash.

Most probably the driver will print again, as the patches are done as
simple as possible, but I cannot test this. Also the probablility that
someone has this printer (Lexmark Color Jetprinter 2050) is low, as it
is from around 20 years ago. No one reported a bug about this crash.