Added driver m2300w
Added the another driver of the group of the unmaintained drivers
which come with their own PPD files (not using PPD files from
foomatic-db).

The driver supports user-added color profiles which can be put into
the

    /var/snap/ghostscript-printer-app/common/m2300w/psfiles

directory.

Otherwise using this driver is as simple as adding a printer and
choosing either auto-selection or the appropriate model from the
driver list.

The supported printers (magicolor 2300W and 2400W) appear both under
Minolta and KONICA MINOLTA in the list (extra PPD files created for
that) as it seems that these printers were still on the market when
Minolta and Konica joined.