Added driver fxlinuxprint
Added another driver of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

We did not give up on this driver as it is for high-end office laser
printers, so many of these printers supported by this 5-year-old
driver are probably still in use.

Fuji Xerox probably stopped with this driver as this is a PDF driver
and nowadays PDF printers use driverless IPP. The driver seems to be
for the pre-driverless-IPP PDF printers.

We add a list of the actually supported printer models to the single
included PPD file as 114 (!) "*Product:" lines by a patch, so that the
user of the Printer Application gets guided to his driver if he has an
appropriate printer. List taken from debian/control file. It seems
that there is no other place which tells about the supported models.

Another patch is added to fix bugs in the filter executables, as they
polled the PPD file from the CUPS daemon instead of using the PPD
environment variable, and did it even wrong (querying the default
printer and not the current printer). There were also some too small
buffers which were probably the reason why this driver was crashing
right away.

The driver also has introduced its own MIME type
"application/vnd.cups-pdfprintfx" which it generates with the
additional filter "pdftopdffx". For this we have added a new
conversion rule to ghostscript-printer-app.c.