Added driver printer-driver-oki
Added another driver of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

This driver is made by OKI, but it is simple, partially PostScript
printer PPDs which do not need a CUPS filter, partially PPD files
using simple shell scripts as filters which use the "epson"
Ghostscript driver (for dot matrix printers) and CUPS' rastertohp
filter (for PCL laser printers), where the scripts arrange PPD option
settings and add PJL if needed.

Adjusted paths in the scripts and corrected a language specification
in a PPD via perl regexp magic, so that all builds correctly and
works.