Added driver pxljr
Added another driver of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

This driver did not need any special treatment to get working in
the Snap. The manipulations on the included PPD are exactly the
same as in the Debian package.