Added driver rastertosag-gdi
Added another driver of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

This driver did not need any special treatment to get working in
the Snap. Only change is the Python 3 adaptation by the patch on the
Debian package.