Added driver splix
Added another driver of the group of the unmaintained drivers which
come with their own PPD files (not using PPD files from foomatic-db).

Here we principally build as in the Debian package but instead of
using the *.drv file we use the included, already built PPD files. We
also set the CUPS filter path so that the pstoqpdl filter finds
gstoraster.

In addition, we let the cups-filters part also install the gstoraster
filter stub ad SpliX (pstoqpdl) needs it.