Added driver brlaser
This driver for Brother laser printers received commits on its GitHub
repository until relatively recently (mid-2020), so it is not
completely sure whether it is still maintained or not.

As it has several changes after its last release and the Debian
package has no patches, this driver is taken from the upstream
repository and not from the Debian packaging GIT as most other
drivers.