Added driver ptouch-driver
This driver for Brother label printers received commits on its GitHub
repository until recently (mid-2021), so one can assume that it is
actually maintained.

As it has some changes after its last release and the Debian package
has no patches, this driver is taken from the upstream repository and
not from the Debian packaging GIT as most other drivers.

It is built the same way as the Debian package, especially the PPD
files are pre-built from the Foomatic XML data.