snapcraft.yaml: Make use of libppd's *.drv support, clean-up
libppd supports driver information files (*.drv) now. So we do not
need to get ppdc from CUPS, where it will go away later and we also do
not need to pre-build PPD files from *.drv files but simply include
the *.drv for on-the-fly generation of the PPD files.

For the *.drv file of cups-filters we still pre-build as we are
removing some of the PPDs generated by the *.drv file.

Also fixed the *.drv files of the c2esp driver. They contain absolute
paths to the CUPS filters. Simply removed them with Perl.