Switch to the PDF-based test page and debugging copies
CUPS uses for lonfer time already the PDF-based test page of
cups-filters. Here the bannertopdf filter adds information about the
job and the printer to a template PDF page which gets scaled to the
page size of the job/the printer.

The bannertopdf() filter function is now adapted to work in a
pappl-retrofit-based Snap and so we could switch over.

This allows creating customized templates with modern graphics
editors, no PostScript programming needed. Also the test page contains
information about printer make/model/driver (the PPD file's NickName
entry), and its is replicated several times when printing it duplex or
with N-up, so that the sheet gets completely filled to see the effect
of the settings of these options.

In addition, we create a debugging copy of what gets sent to the
printer when we are running in debug logging mode ("-o
log-level=debug" or appropriate setting in the web interface). The
copies are put into the spool directory and the file names contain the
print queue's name and the job number. To not clog up the disk the
files get deleted after 24 hours.