c2esp needs adaptation to libcupsfilters 2.x API
c2esp is, according to "apt rdepends libcupsfilters1", the only
printer driver (at least from the ones available in Debian/Ubuntu)
which uses libcupsfilters and so can suffer from the recent API change
on driver-related functions to introduce the log function concept.

And indeed it does not build any more with libcupsfilters of
cups-filters 2.x and so makes the build of the Ghostscript Printer
Application fail.

This commit adds a simple patch to adapt the code of c2esp to the new
API.