Use $CRAFT_PART_INSTALL for install path
When copying the "make" command lines from the log of the Snap build
process for the "ghostscript-printer-app" part for calling them
explicitly in the "override-build", the path resolved from this
variable was copied, not the variable name in the command line. This
is corrected now.

Also removed the explicit mention of the files in the "stage" section
again.